+++ .env.example
@@ -2,3 +2,5 @@ HOST=http://localhost
 PORT=3000
 BASE_URL=http://localhost:3000
 MORPHURL_URL=https://morphurl.herokuapp.com
+MORPHCLOUDS_URL=https://morphclouds.herokuapp.com
+MORPHCLOUDS_TOKEN=
