+++ src/clients/pages/home/Hero/Hero.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/tokens/all';
+@import '../../../../../styles/tokens/all';
 
 .hero {
   padding: $space--sm 0;
