+++ src/clients/pages/home/MorphsearchCopy/MorphsearchCopy.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/tokens/all";
+@import '../../../../../styles/tokens/all';
 
 .morphsearchCopy {
   background: $color-white;
