+++ src/clients/pages/home/ProductList/ProductList.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/tokens/all";
+@import '../../../../../styles/tokens/all';
 
 .productList {
