+++ src/clients/pages/home/ShortenurlCopy/ShortenurlCopy.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/tokens/all';
+@import '../../../../../styles/tokens/all';
 
 .shortenurlCopy {
   background: $color-yellow;
