+++ src/clients/pages/shortenurl/ShorturlForm/ShorturlForm.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/tokens/all";
+@import "../../../../../styles/tokens/all";
 
 .shortenurlForm {
   background: $color-white;
