+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import MetaHead from '../components/Head/Head';
 import HomepageMeta from '../../public/seo/homepage.json';
-import MorphsearchCopy from '../clients/Homepage/MorphsearchCopy/MorphsearchCopy';
-import ShortenurlCopy from '../clients/Homepage/ShortenurlCopy/ShortenurlCopy';
-import Hero from '../clients/Homepage/Hero/Hero';
+import MorphsearchCopy from '../clients/pages/home/MorphsearchCopy/MorphsearchCopy';
+import ShortenurlCopy from '../clients/pages/home/ShortenurlCopy/ShortenurlCopy';
+import Hero from '../clients/pages/home/Hero/Hero';
 
 const Home = () => {
   return (
